requirements.txt
streamlit
pathlib
python-dotenv
google-generativeai
youtube_transcript_api